restructure to search component
