+++ src/styles/sections/header/header.scss
@@ -1,5 +1,4 @@
 @import "../../variables.scss";
-@import "./search.scss";
 @import "./account.scss";
 @import "./toggleNav.scss";
